moved back Rectangle object from the domain package.
