+++ src/main/java/ua/comparison/image/ImageComparisonTools.java
@@ -81,6 +81,29 @@ public class ImageComparisonTools {
     }
 
     /**
+     * Create a {@code Rectangle} object.
+     * @param matrix the matrix of the Conformity pixels.
+     * @param counter the number from marks regions.
+     * @return the {@code Rectangle} object.
+     */
+    public static Rectangle createRectangle( int[][] matrix, int counter ) {
+        Rectangle rect = Rectangle.createDefaultRectangle();
+
+        for ( int y = 0; y < matrix.length; y++ ) {
+            for ( int x = 0; x < matrix[0].length; x++ ) {
+                if ( matrix[y][x] == counter ) {
+                    if ( x < rect.getMinX() ) rect.setMinX( x );
+                    if ( x > rect.getMaxX() ) rect.setMaxX( x );
+
+                    if ( y < rect.getMinY() ) rect.setMinY( y );
+                    if ( y > rect.getMaxY() ) rect.setMaxY( y );
+                }
+            }
+        }
+        return rect;
+    }
+
+    /**
      * Reads image from the provided path.
      * @param path the path where contains image.
      * @return the {@code BufferedImage} object of this specific image.
